added Back button
In the updated code:

The backButtonCircleOutline style is added to give the back button a gray circle outline.
The backIcon style is added to set the color of the back icon to gray.
The smallCircle style is added to create a small gray circle to the right of the billing header.
The backButton, backButtonCircleOutline, backIcon, and smallCircle styles are used in the appropriate components to achieve the desired appearance.